add stats to overview in frontend, fix overview endpoint in backend

+++ API/RequirementAI.Contract/Dto/ProjectQualityOverviewDto.cs
@@ -2,11 +2,11 @@ namespace RequirementAI.Contract.Dto;
 
 public class ProjectQualityOverviewDto
 {
-    public double TotalProjectScore { get; set; }
+    public decimal TotalProjectScore { get; set; }
 
-    public double AveragePersonaScore { get; set; }
-    public double AverageScenarioScore { get; set; }
-    public double AverageUserStoryScore { get; set; }
+    public decimal AveragePersonaScore { get; set; }
+    public decimal AverageScenarioScore { get; set; }
+    public decimal AverageUserStoryScore { get; set; }
 
     public LowestScoreItemDto? LowestPersona { get; set; }
     public LowestScoreItemDto? LowestScenario { get; set; }
